simplified test + test code cleanup
