+++ soda/core/tests/cli/test_return_values.py
@@ -1,12 +1,31 @@
 from textwrap import dedent
+import pytest
 
 from tests.cli.run_cli import run_cli
 from tests.helpers.common_test_tables import customers_test_table
 from tests.helpers.mock_file_system import MockFileSystem
 from tests.helpers.scanner import Scanner
-from tests.conftest import test_data_source, data_source_config_str
+from tests.conftest import test_data_source
 
 
+def get_ds_definition():
+    return dedent(
+        """
+        data_source cli_ds:
+            type: postgres
+            connection:
+                host: localhost
+                username: sodasql
+            database: sodasql
+            schema: public
+        """
+    ).strip()
+
+
+@pytest.mark.skipif(
+    test_data_source != "postgres",
+    reason="Run for postgres only as nothing data source specific is tested..",
+)
 def test_non_existing_files(scanner: Scanner):
     table_name = scanner.ensure_test_table(customers_test_table)
 
@@ -14,7 +33,7 @@ def test_non_existing_files(scanner: Scanner):
         [
             "scan",
             "-d",
-            test_data_source,
+            "cli_ds",
             "-c",
             "non-existing.yml",
             "checks.yml",
@@ -23,13 +42,17 @@ def test_non_existing_files(scanner: Scanner):
     assert result.exit_code == 3
 
 
-def test_ok_with_variable(scanner: Scanner, mock_file_system: MockFileSystem, data_source_config_str: str):
+@pytest.mark.skipif(
+    test_data_source != "postgres",
+    reason="Run for postgres only as nothing data source specific is tested..",
+)
+def test_ok_with_variable(scanner: Scanner, mock_file_system: MockFileSystem):
     table_name = scanner.ensure_test_table(customers_test_table)
 
     user_home_dir = mock_file_system.user_home_dir()
 
     mock_file_system.files = {
-        f"{user_home_dir}/configuration.yml": data_source_config_str,
+        f"{user_home_dir}/configuration.yml": get_ds_definition(),
         f"{user_home_dir}/checks.yml": dedent(
             f"""
                 checks for {table_name}:
@@ -42,7 +65,7 @@ def test_ok_with_variable(scanner: Scanner, mock_file_system: MockFileSystem, da
         [
             "scan",
             "-d",
-            test_data_source,
+            "cli_ds",
             "-c",
             "configuration.yml",
             "-v",
@@ -53,13 +76,17 @@ def test_ok_with_variable(scanner: Scanner, mock_file_system: MockFileSystem, da
     assert result.exit_code == 0
 
 
-def test_fail(scanner: Scanner, mock_file_system: MockFileSystem, data_source_config_str: str):
+@pytest.mark.skipif(
+    test_data_source != "postgres",
+    reason="Run for postgres only as nothing data source specific is tested..",
+)
+def test_fail(scanner: Scanner, mock_file_system: MockFileSystem):
     table_name = scanner.ensure_test_table(customers_test_table)
 
     user_home_dir = mock_file_system.user_home_dir()
 
     mock_file_system.files = {
-        f"{user_home_dir}/configuration.yml": data_source_config_str,
+        f"{user_home_dir}/configuration.yml": get_ds_definition(),
         f"{user_home_dir}/checks.yml": dedent(
             f"""
                 checks for {table_name}:
@@ -72,7 +99,7 @@ def test_fail(scanner: Scanner, mock_file_system: MockFileSystem, data_source_co
         [
             "scan",
             "-d",
-            test_data_source,
+            "cli_ds",
             "-c",
             "configuration.yml",
             "checks.yml",
@@ -81,13 +108,17 @@ def test_fail(scanner: Scanner, mock_file_system: MockFileSystem, data_source_co
     assert result.exit_code == 2
 
 
-def test_warn(scanner: Scanner, mock_file_system: MockFileSystem, data_source_config_str: str):
+@pytest.mark.skipif(
+    test_data_source != "postgres",
+    reason="Run for postgres only as nothing data source specific is tested..",
+)
+def test_warn(scanner: Scanner, mock_file_system: MockFileSystem):
     table_name = scanner.ensure_test_table(customers_test_table)
 
     user_home_dir = mock_file_system.user_home_dir()
 
     mock_file_system.files = {
-        f"{user_home_dir}/configuration.yml": data_source_config_str,
+        f"{user_home_dir}/configuration.yml": get_ds_definition(),
         f"{user_home_dir}/checks.yml": dedent(
             f"""
                 checks for {table_name}:
@@ -101,7 +132,7 @@ def test_warn(scanner: Scanner, mock_file_system: MockFileSystem, data_source_co
         [
             "scan",
             "-d",
-            test_data_source,
+            "cli_ds",
             "-c",
             "configuration.yml",
             "checks.yml",
